# [Perf] Merge negated matchers on same label into a single regexp matcher