This is a first naive attempt to fix https://github.com/prometheus/prometheus/issues/14619

However, this only shows worse performance:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M3 Pro
                                                                   │     main     │                  join                  │
                                                                   │    sec/op    │    sec/op      vs base                 │
Querier/Head/PostingsForMatchers/RUN_i!~"1",...,i!~"55"-12           923.3n ± 10%   2626.0n ± 19%  +184.41% (p=0.000 n=10)
Querier/Head/PostingsForMatchers/RUN_i!~"1.+",i!~".*2.*",i!="2"-12   8.078m ± 22%    8.881m ± 14%         ~ (p=0.481 n=10)
geomean                                                              86.36µ          152.7µ         +76.83%

                                                                   │     main     │                 join                  │
                                                                   │     B/op     │     B/op      vs base                 │
Querier/Head/PostingsForMatchers/RUN_i!~"1",...,i!~"55"-12             560.0 ± 0%    3320.0 ± 0%  +492.86% (p=0.000 n=10)
Querier/Head/PostingsForMatchers/RUN_i!~"1.+",i!~".*2.*",i!="2"-12   8.633Mi ± 0%   6.723Mi ± 0%   -22.12% (p=0.000 n=10)
geomean                                                              69.53Ki        149.4Ki       +114.87%

                                                                   │    main    │                 join                 │
                                                                   │ allocs/op  │  allocs/op   vs base                 │
Querier/Head/PostingsForMatchers/RUN_i!~"1",...,i!~"55"-12           23.00 ± 0%    56.00 ± 0%  +143.48% (p=0.000 n=10)
Querier/Head/PostingsForMatchers/RUN_i!~"1.+",i!~".*2.*",i!="2"-12   25.00 ± 0%   170.00 ± 0%  +580.00% (p=0.000 n=10)
geomean                                                              23.98         97.57       +306.90%
```

I'll leave this PR here for a few days while I think about it, unless someone has a clue of what's going so wrong.

<!--
    Please give your PR a title in the form "area: short description".  For example "tsdb: reduce disk usage by 95%"

    If your PR is to fix an issue, put "Fixes #issue-number" in the description.

    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --signoff flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->